feat: create auth controller
